chore(restructureApiFiles): restructure files and refactor code
resturcture the files in challenge two
refactor the code
rewrite the tests

[Delivers #162444068]
