+++ app.js
@@ -3,6 +3,5 @@ const express = require("express");
 const routes = require("./routes");
 
 const app = express();
-// app.use("/api/v1/users", routes.user);
 app.use("/api/v1/redflags", routes.redflag);
 module.exports = app;
