--- resources/redflag/all.js
@@ -1,9 +1,0 @@
-const data = require("../../data.json");
-
-module.exports = (req, res) => {
-  const redflag = data.redflag;
-  res.status(200).json({
-    status: 200,
-    redflag
-  });
-}
