+++ routes.js
@@ -1,5 +1,4 @@
 module.exports = {
-  // user: require("./resources/user"),
   // eslint-disable-next-line global-require
-  redflag: require("./resources/redflag")
+  redflag: require("./server/redflagController")
 };
