+++ test/redflag.spec.js
@@ -14,7 +14,8 @@ describe("Red-Flag API Tests", () => {
         .get("/api/v1/redflags")
         .set("Accept", "application/json")
         .expect("Content-Type", /json/)
-        .expect(200, done);
+        .expect(200)
+      done();
     })
   });
   describe("/GET a single red-flag record", () => {
@@ -23,7 +24,8 @@ describe("Red-Flag API Tests", () => {
         .get("/api/v1/redflags/1")
         .set("Accept", "application/json")
         .expect("Content-Type", /json/)
-        .expect(200, done);
+        .expect(200)
+      done();
     })
   });
   describe("/POST a red-flag record", () => {
@@ -43,7 +45,8 @@ describe("Red-Flag API Tests", () => {
         })
         .set("Accept", "application/json")
         .expect("Content-Type", /json/)
-        .expect(200, done);
+        .expect(200)
+      done();
     })
   });
   describe("/DELETE a red-flag record", () => {
@@ -52,7 +55,8 @@ describe("Red-Flag API Tests", () => {
         .delete("/api/v1/redflags/2")
         .set("Accept", "application/json")
         .expect("Content-Type", /json/)
-        .expect(200, done);
+        .expect(200)
+      done();
     })
   });
   describe("/UPDATE the location of a red-flag record", () => {
@@ -72,7 +76,8 @@ describe("Red-Flag API Tests", () => {
         })
         .set("Accept", "application/json")
         .expect("Content-Type", /json/)
-        .expect(200, done);
+        .expect(200)
+      done();
     })
   });
   describe("/UPDATE the comment of a red-flag record", () => {
@@ -92,7 +97,8 @@ describe("Red-Flag API Tests", () => {
         })
         .set("Accept", "application/json")
         .expect("Content-Type", /json/)
-        .expect(200, done);
+        .expect(200)
+      done();
     })
   });
-});
+})
